Use custom OAuthEvents class
Use a custom OAuthEvents class and move the loopback implementation there.
